requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiohttp>=3.9.0
jinja2>=3.1.0
python-dotenv>=0.19.0
requests>=2.26.0
aiofiles>=0.7.0
email-validator>=2.1.0.post1
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0 